Add ListFruit methods and change order
