+++ app/src/main/java/com/example/jetpackcomposeinstagram/data/Retrofit.kt
@@ -3,7 +3,7 @@ package com.example.jetpackcomposeinstagram.data
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
-object RetrofitHelper {
+class Retrofit {
     fun getRetrofit(): Retrofit {
         return Retrofit.Builder()
             .baseUrl("https://run.mocky.io")
