+++ app/src/main/java/com/example/jetpackcomposeinstagram/data/listfruititems/remote/model/Constants.kt
@@ -1,0 +1,5 @@
+package com.example.jetpackcomposeinstagram.data.listfruititems.remote.model
+
+internal object Constants {
+    const val LISTAFRUTAS = "list_fruit"
+}
