+++ app/src/main/java/com/example/jetpackcomposeinstagram/data/listfruititems/remote/model/RemoteList.kt
@@ -1,0 +1,8 @@
+package com.example.jetpackcomposeinstagram.data.listfruititems.remote.model
+
+import com.example.jetpackcomposeinstagram.data.listfruititems.remote.model.Constants.LISTAFRUTAS
+import com.google.gson.annotations.SerializedName
+
+data class RemoteList(
+    @SerializedName(LISTAFRUTAS) val listfruit: List<RemoteListFruit?>?
+)
