+++ app/src/main/java/com/example/jetpackcomposeinstagram/data/login/LoginRepository.kt
@@ -1,0 +1,16 @@
+package com.example.jetpackcomposeinstagram.data.login
+
+import com.example.jetpackcomposeinstagram.data.login.remote.LoginRemoteImpl
+import com.example.jetpackcomposeinstagram.data.login.remote.model.RemoteLogin
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+
+class LoginRepository {
+
+    private val api = LoginRemoteImpl()
+
+    fun doLogin(): Flow<RemoteLogin> = flow {
+        val response = api.doLogin()
+        emit(response as RemoteLogin)
+    }
+}
