--- app/src/main/java/com/example/jetpackcomposeinstagram/data/remote/login/LoginRepository.kt
@@ -1,15 +1,0 @@
-package com.example.jetpackcomposeinstagram.data.remote.login
-
-import com.example.jetpackcomposeinstagram.data.remote.login.models.RemoteLogin
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-
-class LoginRepository {
-
-    private val api = LoginService()
-
-    fun doLogin(): Flow<RemoteLogin> = flow {
-        val response = api.doLogin()
-        emit(response as RemoteLogin)
-    }
-}
