--- app/src/main/java/com/example/jetpackcomposeinstagram/data/remote/login/LoginService.kt
@@ -1,16 +1,0 @@
-package com.example.jetpackcomposeinstagram.data.remote.login
-
-import com.example.jetpackcomposeinstagram.data.RetrofitHelper
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-
-class LoginService {
-    private val retrofit = RetrofitHelper.getRetrofit()
-
-    suspend fun doLogin():Any {
-        return withContext(Dispatchers.IO){
-           retrofit.create(LoginClient::class.java).doLogin()
-        }
-    }
-
-}
